Refactor test_service_task_cancellation_dag_order and run it against trio
